requirements.txt
numpy
pandas
scikit-image
scikit-learn
opencv-python
imagecodecs
numcodecs
tifffile
zarr
omero-py
ome-types
rsa
tqdm
psutil
PyYAML
Pillow
imageio
matplotlib